requirements.txt
pathway==0.8.2
streamlit==1.31.0
python-dotenv==1.0.1
tiktoken
requests
litellm==1.17.3
unstructured[all-docs]==0.10.28
mpmath==1.3.0
